Update patch set 1
Patch Set 1:

Graphadmin is still hardcoded as 1.0 (it was introduced in casablanca) . @Harish, should we move graphadmin up to 1.4 so it follows the convention of the rest of the docker images?  What do we do for the casablanca release, some of the images have patch revs (i.e. 1.3.1 vs. 1.3.2)

Patch-set: 1
Reviewer: Gerrit User 33 <33@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0